requirements.txt
langchain
langchain-community
langchain-openai
faiss-cpu
pypdf
sentence-transformers
flashrank
streamlit
watchdog
langchain-core
langchain-huggingface
python-dotenv
transformers
torch